Test data and scripts are removed from the repository. If one wants to run the tests, the NOA_TEST_DATA env variable should be defined and will be used by noa_test to find the noa-data repository.
Bug fix: CUDA::Math: reduceAdd and reduceMean had the TWO_BY_TWO options. This is now removed since this optimization is not correct for these functions.

Bug fix: In TestCPUIndexes.cpp, idx_min and idx_max could be the same for a given batch. This was generating unexpected data. This is now fixed.

All tests are passing.