Add affine transforms for arrays.
On the CPU backend, Interpolate.h contains the interpolation functions, for
all supported interpolation methods. It also contains the bspline prefilter.
Interpolator.h contains the 1D/2D/3D interpolators, which have a single get
template function. This function takes the coordinates, collects the sample
values which are within the interpolation window. OOB values are handled
according to the `BorderMode` which is a template argument. Then, these
values and the fraction(s) are send to the interpolation functions in
Interpolate.h. Note that the interpolation is always performed, even if all
the values are OOB. For `BORDER_ZERO` and `BORDER_VALUE`, we could detect
when all values are OOB and skip the interpolation, but this is an additional
thing to compute and since in most cases we are at least partially in bound,
I am not sure this will be more efficient. One inconvenient about the get
template function is that InterpMode and BorderMode are template arguments.
This is of course for obvious efficiency reasons, but it makes the launch
quite annoying when these arguments are only known at runtime.

On the CUDA backend, the border/addressing and interpolation is done by the
textures. Interpolate.h contains the tex*D template functions which add
support for `INTERP_COSINE` and `INTERP_CUBIC_BSPLINE` (both are based on
the linear filtering of the textures) while keeping a shared API for all
`InterpMode`. Complex values are supported using the float2 descriptor to
keep things efficient.

Still working on rotate/scale/shift for FFT outputs. This should be added soon.

Tags: [feature]
Compile: yes
Tests: unknown, not added