Refactor the streams.
The unified API has a current `Device` and a current `Stream` for each device.
This state is private to a (host) thread. By default, the current device is the CPU
and the current stream is `Stream::DEFAULT`, i.e. the main thread. The default current
stream for the GPUs is also the `Stream::DEFAULT`.

Creating a `Stream` without any argument creates a "proper" stream, i.e. `Stream::ASYNC`
on the current `Device`.

* `cuda::Device` default ctor retrieves the current device (which by default is ID=0).
  Before it was always using the ID=0 device.
* Default stream ctors create uses the `ASYNC` mode.
* `cpu::Stream` stores its logic (~192 bytes) on the heap. Thus, the stream can be moved now.
* Add unified `Stream`. Use type erasure to contain a `cuda::Stream` or a `cpu::Stream`.
  Use small buffer optimization to store the underlying stream.

Tags: [refactor]
Compile: no
Tests: no