Switch back to OpenBLAS
So... Eigen is great but it's not what I'm looking for simply because compilation
times are crazy (we are talking 30sec-1min for a single BDSVD instantiation).
Moreover, CLion (and its clang tools) really struggle to analyse the code in real
time since there's so much template code from Eigen included in the source file.
Maybe I'll revisit Eigen in the future, but for now, I'll stick with OpenBLAS for
everything related to BLAS and LAPACK operations.

* Build looks for OpenBLAS already installed in the system. The search can be
  guided using `CMAKE_PREFIX_PATH` in case OpenBLAS is installed in an unusual
  location. In the future, I want to build OpenBLAS internally so that we can
  easily make sure it was compiled with the same compiler and includes LAPACKE.

Tags: [cmake]
Compile: no
Tests: no