Update `cpu::fft` namespace to be a bit more flexible with F-contiguity.
The issue with non-redundant FFT is that we often assume the redundancy is along
the rows and when computing the `.fft()` shape, we divide the row dimension. For
now, keep it that way but support edge case of column-vectors. In the future,
maybe add support for 'F' layout for the `.fft()` method on shapes.

When dealing with the full transform, we can swap the dimensions to be more
cache-friendly.

Tags: [refactor]
Compile: yes
Tests: no