Update `cpu::memory` namespace to be a bit more flexible with F-contiguity.
* Return step in `linspace` even with async overload.
* Add multithreading (one thread per subregion) on `extract()` and `insert()`.
* Remove out-of-place permutation implementation and replace with `copy()`.

When there are multiple arrays, the current refactoring takes the output array
and find the reordering to ensure we are about to loop through that array in
the rightmost order. This reordering might make things worse for the other
arrays, but in most cases it should be fine.

Tags: [refactor][feature][doc]
Compile: yes
Tests: no