Refactor Fourier projections
* Add quaternion support for the rotation
* Operators are much more flexible on the supported types, but have much tighter checks.
* Add optional weights directly in the operators.

Note: the backend will be refactored in the coming months. For now, I still need to write the wrapper for each backend, which is extremely annoying. The refactoring will remove most of the logic and the unified interface will deal with the operator creation and call to ewise/iwise etc.