WIP - save, moving back to France (one day I'll learn git I swear)
The big C++20 refactoring is done, the frontend is now handle most of the logic, making backend easier to maintain or easier to add new backends, but it is also simpler to add new features to the library, since in most cases we don't need to interact with the backends and can right "normal" code using the unified API. yay!

Now, tests need to be updated, which shouldn't take long and force me to polish and do a final proofread.