Add TIFF support back again...
libtiff has a very outdated and very limited API. I still managed to enable multithreading reads from the same file stream, and it scales well with the number of threads, especially when decompression is involved. Couldn't do anything for writing though...

ImageFile has been improved, especially the file encoding detection. Now when reading a file, the encoder is not selected on the file extension. Instead, we ask each encoder if its recognizes the file.

BinaryFile now allows postponing the file resizing when creating a new file for cases when we need to let the encoder resize the stream. In this case, memory mapping is not allowed.