Remap 'noa::Remap' to 'noa::fft::Layout'
we often use the namespace alias `nf::Layout`, which is as long as noa::Remap

we often use the implicit string conversion e.g. ‘h2h’, especially in user code

now it makes sense since we can specify the single layout: ‘h’ etc.

in the future we may add `noa::Layout` (Stridded, Contiguous, Compressed, Sparse, etc.), so this needs to be in `noa::fft`